// frontend/src/components/common/UI/Buttons/base/types.ts
export type ButtonSizeValue = "sm" | "md" | "lg";
export enum ButtonSize {
  SMALL = "sm",
  MEDIUM = "md",
  LARGE = "lg",
}

export type ButtonColorValue = "black" | "gray" | "lightGray";
export enum ButtonColor {
  BLACK = "black",
  GRAY = "gray",
  LIGHTGRAY = "lightGray",
}

export type ButtonShapeValue = "square" | "round";
export enum ButtonShape {
  SQUARE = "square",
  ROUND = "round",
}